refactor useEmpList into a container so that it doesn't refresh every reroute
